feat(next): relax root dynamic rendering for auth pages
Remove force-dynamic from the root layout and isolate useSearchParams logic into client content components wrapped in Suspense server pages so auth routes can prerender safely.
